♻️ 関数をファイル分けした&WRONG_SCOPE_COMMENT以外のnormエラーをなくした
