+++ src/debug/debug_print_data.c
@@ -6,39 +6,35 @@
 /*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/07 21:31:41 by miyuu             #+#    #+#             */
-/*   Updated: 2025/05/10 17:05:44 by miyuu            ###   ########.fr       */
+/*   Updated: 2025/05/10 20:45:03 by miyuu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-// void	debug_print_data(t_data *data)
-// {
-// 	int	y;
-// 	int	x;
+void	debug_print_data(t_data *data)
+{
+	size_t	y;
 
-// 	printf("北 path: %s\n", data->texture->no_img);
-// 	printf("南 path: %s\n", data->texture->so_img);
-// 	printf("西 path: %s\n", data->texture->we_img);
-// 	printf("東 path: %s\n", data->texture->ea_img);
-// 	// printf("f_color: %d\n", data->texture->f_color);
-// 	// printf("c_color: %d\n", data->texture->c_color);
-// 	printf("f_color: %s\n", data->texture->f_color);
-// 	printf("c_color: %s\n", data->texture->c_color);
-// 	y = 0;
-// 	x = 0;
-// 	while (data->map[y] != NULL)
-// 	{
-// 		x = 0;
-// 		while (data->map[y][x] != '\0' && data->map[y][x] != '\n')
-// 		{
-// 			if (y == data->player.y && x == data->player.x)
-// 				printf("\x1b[31m%c\x1b[39m", data->map[y][x]);
-// 			else
-// 				printf("%c", data->map[y][x]);
-// 			x++;
-// 		}
-// 		printf("\n");
-// 		y++;
-// 	}
-// }
+	printf("----------- パース後 -------------\n");
+	printf("MLX pointer: %p\n", data->mlx);
+	printf("北 texture: %p\n", data->no_img);
+	printf("南 texture: %p\n", data->so_img);
+	printf("西 texture: %p\n", data->we_img);
+	printf("東 texture: %p\n", data->ea_img);
+	printf("床   color: 0x%06X\n", data->f_color);
+	printf("天井 color: 0x%06X\n", data->c_color);
+	printf("Player position: x = %u, y = %u, dir = %c\n", \
+	data->player.x, data->player.y, data->player.dir);
+	y = -1;
+	printf("以下、mapデータ: ");
+	if (!data->map)
+		printf("なし(null)\n");
+	else
+	{
+		printf("\n");
+		while (data->map && data->map[++y] != NULL)
+			printf("%s\n", data->map[y]);
+	}
+	printf("---------------------------------\n\n");
+}
